+++ modules/Attendance/TeacherCompletion.php
@@ -27,7 +27,7 @@ foreach ( (array) $categories_RET as $category )
 
 $category_select .= '</select>';
 
-$periods_RET = DBGet( "SELECT sp.PERIOD_ID,sp.TITLE
+$periods_RET = DBGet( "SELECT sp.PERIOD_ID,sp.TITLE,COALESCE(sp.SHORT_NAME,sp.TITLE) AS SHORT_TITLE
 	FROM school_periods sp
 	WHERE sp.SCHOOL_ID='" . UserSchool() . "'
 	AND sp.SYEAR='" . UserSyear() . "'
@@ -58,7 +58,7 @@ echo '</form>';
 
 if ( SchoolInfo( 'NUMBER_DAYS_ROTATION' ) !== null )
 {
-	$sql = "SELECT s.STAFF_ID," . DisplayNameSQL( 's' ) . " AS FULL_NAME,
+	$sql = "SELECT s.STAFF_ID," . DisplayNameSQL( 's' ) . " AS FULL_NAME,s.ROLLOVER_ID,
 	sp.TITLE,cpsp.PERIOD_ID,cp.TITLE AS CP_TITLE,
 	(SELECT 'Y'
 		FROM attendance_completed ac
@@ -164,9 +164,17 @@ if ( ! isset( $_REQUEST['period'] )
 
 	$columns = [ 'FULL_NAME' => _( 'Teacher' ) ];
 
+	$period_title_column = 'TITLE';
+
+	if ( count( $periods_RET ) > 10 )
+	{
+		// Use Period's Short Name when > 10 columns in the list.
+		$period_title_column = 'SHORT_TITLE';
+	}
+
 	foreach ( (array) $periods_RET as $id => $period )
 	{
-		$columns[$id] = $period[1]['TITLE'];
+		$columns[$id] = $period[1][$period_title_column];
 	}
 
 	ListOutput( $staff_RET, $columns, 'Teacher who takes attendance', 'Teachers who take attendance' );